Create the 'StarIcon' component.
